+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 This project adheres to [Semantic Versioning](http://semver.org/). All notable changes will be documented in this file.
 
 ## [Unreleased]
-Nothing yet.
+### Added
+- [#311](https://github.com/Kashoo/synctos/issues/311): Case insensitive equality constraint for strings
 
 ## [2.4.0] - 2018-04-24
 ### Added